+++ SharedAssemblyInfo.cs
@@ -12,4 +12,4 @@ using System.Reflection;
 
 [assembly: AssemblyVersion("8.4.4")]
 [assembly: AssemblyFileVersion("8.4.4")]
-[assembly: AssemblyInformationalVersion("8.4.4-rc1")]
+[assembly: AssemblyInformationalVersion("8.4.4-DataSubTypeFix")]